Do not pan if element's scrollbar is grabbed
== Firefox ==
event.originalTarget is a Mozilla-specific property that refers to
the original event target before any retargetings. This element
could be a XUL element. Access to any properties of this object is
forbidden for normal web content.

This characteristic is used to make sure that clicking on the
scrollbar doesn't activate the panning behavior.

== Opera Presto, Safari, IE ==
To make sure that dragging the scrollbar does not trigger panning,
an onscroll element was added. When this event is triggered before
mousemove, the code assumes that a scrollbar was moved.
Side-effect: When the user uses the cursor or scrollwheel to scroll
the element without having moved the mouse, the pan action is also
cancelled. (this is a very minor problem, a very small edge-case)

== Chrome ==
Already behaved as expected before this commit: grab and drag the
scrollbar did not activate panning.

Successfully verified the correct behavior using demo.html in:
- IE9 (+classList.js polyfill from eligrey), 11
- Opera 11.60, 12 (Presto)
- Chromium 20, 28
- Firefox 11, 23
- Safari 5.1, 6

Verification method:
1. Grab the area and move the mouse around --> should pan smoothly
   Release mouse --> Pan should end
2. Grab scrollbar and move mouse around --> scrollbar should move.
   Release mouse --> Scrollbar should sit in its final position,
   and panning should not occur.