Addresses #1
In Presto and Webkit-based browsers (Chrome, Opera 15+ and Safari),
the cursor style propagates to the scrollbars of an element.

Having the wrong cursor on the scrollbar is very confusing to users,
so I have removed the default grab cursor from the main element.
The grab cursor still shows up on all child elements of the container
though. When panning is activated after grabbing the element, the
cursor is still changed.