Performance improvement for huge document trees
Previously, the cursor was changed by toggling a class name.
In complex and huge documents, this simple action hurt the
performance of the page.
According to the profile (http://stackoverflow.com/q/21392844),
recalculating the styles after toggling the class name took
almost 200ms in a document with 7.7k nodes.

Since this performance issue is unacceptable, I've modified the
method. From now on, moving the cursor during a drag operation
results in a transparent overlay to be added to the document,
which results in the desired effect: a "grabbing" cursor.